Move extention to keymapping, return bool
